requirements.txt
chess
matplotlib
mltu
numpy==1.26.4
opencv_python
pandas
PyYAML
torch==2.5.1
torchsummaryX
tqdm
onnx
tensorboard
timm
transformers
jiwer
evaluate

